DocumentsReaderServer/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
Pillow
requests
pyodbc
python-multipart